Fix Formey (1773) date label
Formey (1773 [1765]) was labelled as Formey (1773) as an edition of the text was published in 1773. Yet, the text had actually been published (or at least was available to read in some form) in 1765. This created potential confusion as at least one text (Borrelly 1768) referred to Formey (1773 [1765]) before 1773. Hence the fix (replacement of Formey (1773) by Formey (1773 [1765]).